HPCC-15130 Reduce jlib dependencies
Move NUMA dependency to thorhelper
