+++ cmake_modules/commonSetup.cmake
@@ -50,7 +50,7 @@ IF ("${COMMONSETUP_DONE}" STREQUAL "")
   option(CLIENTTOOLS_ONLY "Enable the building of Client Tools only." OFF)
   option(INCLUDE_PLUGINS "Enable the building of platform and all plugins for testing purposes" OFF)
   option(PLUGIN "Enable building of a plugin" OFF)
-  option(USE_SHLIBDEPS "Enable the use of dpkg-shlibdeps on ubuntu packagin" OFF)
+  option(USE_SHLIBDEPS "Enable the use of dpkg-shlibdeps on ubuntu packaging" OFF)
 
   if (APPLE OR WIN32)
     option(USE_BINUTILS "Enable use of binutils to embed workunit info into shared objects" OFF)
@@ -281,14 +281,6 @@ IF ("${COMMONSETUP_DONE}" STREQUAL "")
   IF (NOT THREADS_FOUND)
     message(FATAL_ERROR "No threading support found")
   ENDIF()
-
-  if (USE_NUMA)
-    find_package(NUMA)
-    add_definitions (-D_USE_NUMA)
-    if (NOT NUMA_FOUND)
-      message(FATAL_ERROR "Support for numa not found")
-    endif()
-  endif()
 
   if ("${CMAKE_CXX_COMPILER_ID}" MATCHES "Clang")
    set (CMAKE_COMPILER_IS_CLANGXX 1)
@@ -843,6 +835,14 @@ IF ("${COMMONSETUP_DONE}" STREQUAL "")
       else()
         add_definitions (-D_NO_APR)
       endif(USE_APR)
+
+      if (USE_NUMA)
+        find_package(NUMA)
+        add_definitions (-D_USE_NUMA)
+        if (NOT NUMA_FOUND)
+          message(FATAL_ERROR "NUMA requested but package not found")
+        endif()
+      endif()
 
       if(USE_TBB)
         find_package(TBB)
